Release 0.9.4
There were 3 main changes made in this version: the "DOM"-estication of the map 
entities, revamping of the printing system to use the event-delegation model, 
and renaming of the server frame listener and support classes. Also, the 
included copy of OpenXML has been updated to version 1.0.6, and javax.vecmath to 
1.1-1.0. 

The "DOM"-estication of the map entities means that all the constructors for map 
entity classes (such as weapons, ammoboxes, doors, etc), now take DOM Element 
objects as an argument (instead of an array of Strings, which was the old way of 
doing things). There were -so- many classes involved in this change that I won't 
bother trying to list them all. For the most part, the only change was in the 
declaration of the constructor - since the utility methods for reading spawn 
args hide most of the differences between the arg styles. 

The printing system was quite a mess before, with printing being handled in 
several places such as the Player class, the CTF team class, and the Game class. 
Now, everything is unified with a single set of PrintEvent, PrintListener, and 
PrintSupport classes. Player chats, team chats, debugging output, and localized 
announcements all now come through a single mechanism, that can easily be 
listened to, filtered, and even modified (for spam supression, bad word 
filtering, logging, telnet servers, etc). 

The changing of the server frame listener classes, and alteration of the method 
names in the Game class for adding and removing listeners also caused many many 
classes to be slightly changed. But that was pretty much simple search/replace 
name change type stuff, and did't actually change the way things work. So for 
most of those changes I won't bother actually listing them here either.