Release v0.9.6
Two (only two?) big changes this time...the Gamelet and GameletManager classes 
have been reworked/simplified/DOM-esticated, and a new "status" facility has 
been added to the q2java.core.Game class. 

The updated Gamelet scheme is basically: gamelet constructors now take a DOM 
document as an argument (instead of just a string), and the DOM document may 
contain data passed to the GameletManager.addGamelet() method and/or data stored 
in a file with the same name as the gamelet classfile except with a .gamelet 
extension instead of .class. 

The GameletManager no longer supports the notion of "dependencies", the gamelets 
should work that sort of thing out amongst themselves. The existing BIG gamelets 
used to be dependant on the q2java.baseq2.BaseQ2 gamelet, but PITA situation was 
resolved by splitting the "gamelet" part of BaseQ2 off into a new gamelet called 
"q2java.baseq2.Deathmatch", and what was left over was no longer a gamelet, but 
instead is now a support class that -real- gamelets like CTF, Paranoia, 
BountyHunters, and the new Deathmatch could call upon. 

Gamelets no longer have an isLevelChangeRequired() method, or an init() method 
that they can have called when a level starts - instead they should just 
register as GameStatusListeners and find out about level starts the normal way. 
As far as waiting for a level change to unload goes, they can be marked that way 
with a <option name="wait for level change to unload"> tag in the corresponding 
.gamelet file. A <option name="avoid multiple instances"> tag marks gamelets 
that should not be loaded multiple times. 

Because of the Gamelet and Gamelet manager changes, all the gamelets included in 
q2java had to be updated to fit in the new scheme...I won't bother listing each 
one individually. 

The new "status" scheme is basically: the Game class creates a DOM Document 
containing just (in XML notation) <status game="Quake2" ip="a.b.c.d" port="xxx" 
map="yyy"/> and updates the mapname as necessary. This document can be fetched 
by calling Game.getDocument("q2java.status"), and updated by any interested 
code. If the status document is updated, 
Game.notifyDocumentChanged("q2java.status") should be called, so a 
GameStatusEvent.GAME_DOCUMENT_CHANGED event will be fired off and interested 
objects will learn about the change. 

That may not sound very interesting, but that's what makes it so powerful. It's 
completely up to gamelet code to decide what to put in the status document, and 
what to do when it's updated. To see the status document feature in action, load 
the barryp.status.PlayerStatus gamelet to have the status document updated to 
show connected players, barryp.status.GameletStatus to have the doc updated to 
show loaded gamelets, and barryp.status.WriteXML to write the document out to a 
file whenever it changes. 

Other gamelets may be written to maintain additional or different info, and do 
other things with the document when it changes, such as fire a copy off to some 
sort of master server through UDP or TCP (as a hypothetical example).