Release v0.9.7
It's been a long time since the previous version, and this one contains many 
minor bugfixes and tweaks, along with some big reorganizations that made things 
cleaner and clearer. 

The damage event system was reworked to also function for non-Player objects. 
Everything that was named *PlayerDamage* was renamed to just *Damage*. All 
fields and methods that were geared to the Player class were changed to work for 
the more general GameObject class. A huge number of classes were affected by 
these changes but for the most part, they were of the search-and-replace type of 
thing that can be fixed in a matter of minutes. The q2java.baseq2.Player class 
in particular had a slew of "PlayerDamage" fields and methods that were for the 
most part simply renamed. 

As part of the damage-system change, all objects had their damage() methods 
replaced with versions that take a single DamageEvent object, instead of a slew 
of parameters. The GameObject class still has a method for initiating damage 
with a slew of parameters, but that simply creates a DamageEvent object and 
passes it on through the new event-system. 

The q2java.baseq2.GenericItem class had a nice overhaul, simplifying its 
internal structure greatly, and removing any potential conflicts with its 
subclasses that wanted to receive ServerFrameListener.runFrame() calls (which 
the old GenericItem class also wanted to do).